7188852: Move implementation of De/Inflater.getBytesRead/Writtten() to java from native
Re-implemented getBytesRead/Writtten() at java level

+++ jdk/src/share/native/java/util/zip/zlib-1.2.5/inflate.c
@@ -1370,7 +1370,7 @@ int ZEXPORT inflateSync(strm)
 z_streamp strm;
 {
     unsigned len;               /* number of bytes to look at or looked at */
-    long long in, out;      /* temporary to save total_in and total_out */
+    unsigned long in, out;      /* temporary to save total_in and total_out */
     unsigned char buf[4];       /* to restore bit buffer to byte string */
     struct inflate_state FAR *state;
